ml-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
scikit-learn==1.3.2
numpy==1.26.2
google-generativeai==0.3.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
